# openapi.yml
openapi: 3.0.0
info:
  title: Trivia API
  description: API for a trivia game
  version: 1.0.0
servers:
  - url: http://localhost:8083
    description: Development server
paths:
...
  /teams:
    post:
      summary: Create a team
      operationId: createTeam
      parameters:
        - name: roomId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Team created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
        "401":
          description: Player not authenticated
        "403":
          description: Only the host can create a team
        "404":
          description: Room not found
        "409":
          description: Cannot create a team during a game

    get:
      summary: Get the teams of a room
      operationId: getTeams
      parameters:
        - name: roomId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Teams retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Team"
        "404":
          description: Room not found
  /teams/{teamId}/players/{playerId}:
    put:
      summary: Assign a player to a team
      operationId: assignPlayerToTeam
      parameters:
        - name: teamId
          in: path
          required: true
          schema:
            type: integer
        - name: playerId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Player assigned to team
        "401":
          description: Player not authenticated
        "403":
          description: Only the host can assign another player to a team
        "404":
          description: Player or team not found
        "409":
          description: Cannot assign a player to a team during a game

    delete:
      summary: Remove a player from a team
      operationId: removePlayerFromTeam
      parameters:
        - name: teamId
          in: path
          required: true
          schema:
            type: integer
        - name: playerId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Player removed from team
        "401":
          description: Player not authenticated
        "403":
          description: Only the host can remove another player from a team
        "404":
          description: Player or team not found
        "409":
          description: Cannot remove a player from a team during a game
  /teams/{id}:
    get:
      summary: Get a team
      operationId: getTeam
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Team retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
        "404":
          description: Team not found

    delete:
      summary: Delete a team
      operationId: deleteTeam
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Team deleted
        "401":
          description: Player not authenticated
        "403":
          description: Only the host can delete a team
        "404":
          description: Team not found
        "409":
          description: Cannot delete a team during a game
...
components:
  schemas:
...
    Team:
      type: object
      properties:
        id:
          type: integer
        roomId:
          type: integer
...